# src/openapi/dist/openapi.yaml
openapi: 3.1.0
info:
  title: mucajey Card API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.mucajey.twicemind.com
    description: Production server
paths:
...
  /edition:
    post:
      summary: >-
        Creates a new edition JSON file with the required metadata and an empty
        card list.
      description: >-
        Creates a new edition JSON file with the required metadata and an empty
        card list.
      tags:
        - Edition
      parameters: []
      responses:
        '200':
          description: Successful response
        '201':
          description: Created
        '400':
          description: Bad Request
        '409':
          description: Conflict
  /edition/{edition}:
    get:
      summary: Retrieve a specific edition file by its identifier.
      description: Retrieve a specific edition file by its identifier.
      tags:
        - Edition
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
    put:
      summary: Update a specific edition file by its identifier.
      description: Update a specific edition file by its identifier.
      tags:
        - Edition
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
    delete:
      summary: Delete a specific edition file by its identifier.
      description: Delete a specific edition file by its identifier.
      tags:
        - Edition
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /year/{year}:
    get:
      summary: Filter cards by year
      description: Filter cards by year
      tags:
        - Cards
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal Server Error
...
  /card/id/{edition}/{id}:
    get:
      summary: Retrieve the card that belongs to a specific edition and identifier.
      description: Retrieve the card that belongs to a specific edition and identifier.
      tags:
        - Card
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
  /card/title/{title}:
    get:
      summary: List all cards whose title exactly matches the requested value.
      description: List all cards whose title exactly matches the requested value.
      tags:
        - Card
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /card/artist/{artist}:
    get:
      summary: List all cards whose artist exactly matches the requested value.
      description: List all cards whose artist exactly matches the requested value.
      tags:
        - Card
      parameters:
        - name: artist
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /card/year/{year}:
    get:
      summary: List all cards from the specified year.
      description: List all cards from the specified year.
      tags:
        - Card
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /card/edition/{edition}:
    get:
      summary: List all cards from the specified edition.
      description: List all cards from the specified edition.
      tags:
        - Card
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /card/genre/{genre}:
    get:
      summary: List all cards from the specified genre.
      description: List all cards from the specified genre.
      tags:
        - Card
      parameters:
        - name: genre
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /card/search/{query}:
    get:
      summary: Search for cards matching the query string.
      description: Search for cards matching the query string.
      tags:
        - Card
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /card:
    post:
      summary: Create a new card.
      description: Create a new card.
      tags:
        - Card
      parameters: []
      responses:
        '200':
          description: Successful response
        '201':
          description: Created
        '400':
          description: Bad Request
  /card/{edition}/{id}:
    patch:
      summary: Update a card by edition and id.
      description: Update a card by edition and id.
      tags:
        - Card
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
    delete:
      summary: Delete a card by edition and id.
      description: Delete a card by edition and id.
      tags:
        - Card
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
...